refactor(pack): refactor line diagnostic
