+++ packages/manager/apps/telecom/src/app/telecom/pack/xdsl/access/linediagnostic/linediagnostic.html
